@llogiq 2016-09-24
Thanks! You never disappoint :smile:

I think I've understood your implementation; evil aside, it should work exceedingly well.

Just for the record, can you tell me your CPU model?