@Veedrac 2016-09-24
I'm running an i7-6700HQ.